Use MacVim in foreground mode.
When using MacVim to write commit messages, I need it to launch in
foreground mode. I will likely be adding it as an editor in `git config
--global` as well in the future since adding the `-f` flag was not
sufficient for me to fix the problem on OSX right now. I could add this
to bootstrap.sh and just ensure OS is OSX.