Restructured directory for easier symlinking and added link.sh
link.sh is taken from tomsquest/dotfiles and has a handy function that
will generate symlinks if one does not already exist for a series of
important dotfiles. As such, I have pulled up vimrc and zshrc so that
they are not contained within zsh/ and vim/ which now can be directly
symlinked to ~/.zsh and ~/.vim.

I will add additional files (some are commented out that came from the
original) as they are ready.

Also, I have removed fish/ since I have decided to stick with zsh and
oh-my-zsh and don't need those customizations anymore.