Fixes bugs found when bootsrapping Mac Mini
This commit fixes several bugs. First, the setdiff function from
cowboy/dotfiles was missing and is used inside of all of the brew
formulas. This is now included, as well as removing erroneous spaces
when creating the lists of brew recipes, adding an environment variable
to ensure brew cask symlinks to /Applications and not ~/Applications,
and fixing the sequencing of some dependencies.